requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.4
pytest==7.4.2
httpx==0.25.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
tenacity==8.2.3
email-validator==2.3.0
